@@ -5,8 +5,8 @@
     "start": "NODE_ENV=production node server/server.js"
   },
   "dependencies": {
-    "@turf/transform-rotate": "^4.4.0",
-    "@turf/turf": "^4.4.0",
+    "@turf/transform-rotate": "^5.1.5",
+    "@turf/turf": "^5.1.6",
     "axios": "^0.16.1",
     "body-parser": "^1.17.1",
     "compression": "^1.6.2",
@@ -26,12 +26,12 @@
     "mapbox-gl": "^0.37.0",
     "mapbox.js": "^3.1.1",
     "mongodb": "^2.2.26",
-    "next": "^3.2.3",
+    "next": "^5.1.0",
     "next-redux-wrapper": "^1.1.2",
-    "prop-types": "^15.5.8",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-ga": "^2.2.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "react-ga": "^2.5.7",
     "react-mapbox-gl": "^1.11.0",
     "react-redux": "^5.0.4",
     "redis": "^2.7.1",
