@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 
 const IllustrationRocks = '/static/icons/Illustration_Rocks.svg';
@@ -9,8 +10,8 @@ import * as METRICS from '../../shared/style/metrics';
 class FreedUp extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
 
   static propTypes = {
-    amount: React.PropTypes.number,
-    intl: React.PropTypes.object,
+    amount: PropTypes.number,
+    intl: PropTypes.object,
   }
 
   render = () => (
