@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 
 // Images
@@ -12,8 +13,8 @@ import * as COLORS from '../../shared/style/colors';
 class YourGuessArrow extends React.PureComponent {
 
   static propTypes = {
-    data: React.PropTypes.number,
-    area: React.PropTypes.bool,
+    data: PropTypes.number,
+    area: PropTypes.bool,
   }
 
   render() {
