@@ -1,4 +1,4 @@
-import turf from '@turf/turf';
+import * as turf from '@turf/turf';
 
 export const unfold = function () {
   function geoJsonStreetAnimation(originalStreet, coiledStreet, originLatLon, progressUnfold, progressRestitch) {
